#8 Move file to Editor folder
