+++ src/pages/DocPage.tsx
@@ -3,7 +3,7 @@ import { RouteComponentProps } from 'react-router';
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 
 import NavBar from '../components/NavBar';
-import CodeEditor from '../components/CodeEditor';
+import CodeEditor from '../components/Editor/CodeEditor';
 
 type DocPageProps = {
   docKey: string
